Update temperature from Double to Float

+++ app/src/main/java/com/twitter/challenge/data/WeatherEntity.kt
@@ -2,7 +2,7 @@ package com.twitter.challenge.data
 
 data class WeatherEntity(
         val locationName: String,
-        val tempCelsius: Double,
+        val tempCelsius: Float,
         val windSpeed: Double,
         val cloudiness: Int
 )
