+++ app/src/main/java/com/twitter/challenge/model/Weather.kt
@@ -2,8 +2,8 @@ package com.twitter.challenge.model
 
 data class Weather (
     val locationName: String,
-    val tempCelsius: Double,
-    val tempFahrenheit: Double,
+    val tempCelsius: Float,
+    val tempFahrenheit: Float,
     val windSpeed: Double,
     val cloudiness: Int
 )
